Renamed inspector_tree_flutter.dart to inspector_tree_controller.dart
