+++ packages/devtools_app/lib/src/inspector/inspector_controller.dart
@@ -35,7 +35,7 @@ import 'inspector_screen.dart';
 import 'inspector_service.dart';
 import 'inspector_text_styles.dart' as inspector_text_styles;
 import 'inspector_tree.dart';
-import 'inspector_tree_flutter.dart';
+import 'inspector_tree_controller.dart';
 
 const inspectorRefQueryParam = 'inspectorRef';
