+++ packages/devtools_app/lib/src/inspector/inspector_tree_controller.dart
